+++ src/converter/Extra.scala
@@ -17,38 +17,17 @@
  along with Configrity.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-package org.streum.configrity
+package org.streum.configrity.converter
 
-import converter.ValueConverter
 
-/** Monadic Reader */
-trait Reader[A] {
+trait Extra {
+  
+  // File
 
-  /** Produce a value from a Configuration */
-  def apply( c: Configuration ): A
+  // Color
 
-  def map[B]( f: A => B ) = {
-    val parent = this
-    new Reader[B] {
-      def apply( c: Configuration ) = f( parent(c) )
-    }
-  }
-  def flatMap[B]( f: A => Reader[B] ) = {
-    val parent = this
-    new Reader[B] {
-      def apply( c: Configuration ) = f( parent(c) )(c)
-    }
-  }
+  // URL
 
-}
+  // URI
 
-case class ConfigurationReader[A: ValueConverter](
-  key: String, 
-  default: Option[A]
-) extends Reader[A] {
-  def apply( c: Configuration ) = if( default.isDefined ) {
-    c.get[A](key).getOrElse( default.get )
-  } else {
-    c[A](key)
-  }
 }